// src/types/v46.ts
export interface MarketBonds {
    creation: (Bond | undefined)
    oracle: (Bond | undefined)
    outsider: (Bond | undefined)
}
// ...
export interface Bond {
    who: Uint8Array
    value: bigint
    isSettled: boolean
}